change: change content index to fvn-hash

+++ ip4_set.go
@@ -2,7 +2,7 @@ package main
 
 type Ip4Set map[uint32]ArrayIntSet
 
-func (a *Ip4Set) Delete(ip uint32, id int32) {
+func (a *Ip4Set) Delete(ip uint32, id uint64) {
 	if v, ok := (*a)[ip]; ok {
 		v = v.Del(id)
 		if len(v) == 0 {
@@ -13,7 +13,7 @@ func (a *Ip4Set) Delete(ip uint32, id int32) {
 	}
 }
 
-func (a *Ip4Set) Add(ip uint32, id int32) {
+func (a *Ip4Set) Add(ip uint32, id uint64) {
 	if v, ok := (*a)[ip]; !ok {
 		v = make(ArrayIntSet, 0, 1)
 		(*a)[ip] = v.Add(id)
